rework
adding better effects and making it faster and smoother
